removed the heatmaps functions to put them in heatmap_helpers
